// security_frontend/security_app/src/app/model/revokDto.ts
export enum RevocationReason {
  UNSPECIFIED = 'UNSPECIFIED',
  KEY_COMPROMISE = 'KEY_COMPROMISE',
  CA_COMPROMISE = 'CA_COMPROMISE',
  AFFILIATION_CHANGED = 'AFFILIATION_CHANGED',
  SUPERSEDED = 'SUPERSEDED',
  CESSATION_OF_OPERATION = 'CESSATION_OF_OPERATION',
  CERTIFICATE_HOLD = 'CERTIFICATE_HOLD',
  REMOVE_FROM_CRL = 'REMOVE_FROM_CRL',
  PRIVILEGE_WITHDRAWN = 'PRIVILEGE_WITHDRAWN',
  AA_COMPROMISE = 'AA_COMPROMISE'
}

export interface RevokeCertificateDTO {
  serialNumber: string;
  reason: RevocationReason;
  comment: string;
}